+++ tests/e2e/role/index.js
@@ -1,5 +1,5 @@
 const { describe } = require("mocha");
 
-describe("E2E - 🃏 Roles tests", () => {
+describe("E2E - 🃏 Role tests", () => {
     require("./get-roles.test");
 });